requirements/dev.txt
Django==1.5.1
#PIL==1.1.7
#Pillow==2.1.0
South==0.8.1
django-campaign==0.2.beta
django-debug-toolbar==0.9.4
django-filebrowser==3.5.2
django-grappelli==2.4.5
django-mptt==0.5.5
django-pagination==1.0.7
django-photologue==2.6.1
django-reversion==1.7.1
django-simple-news==1.1
django-tagging==0.3.1
django-tinymce==1.5.1
sorl-thumbnail==11.12
wsgiref==0.1.2
django-moderation
